factor.md
In this code, the user is prompted to enter a number using the input() function, and the input is converted to an integer using the int() function.

The factorial is initialized as 1. Then, a for loop is used to iterate from 1 to the given number (inclusive). In each iteration, the value of i is multiplied with the current value of the factorial variable using the *= operator.